walk: fix cancellation test synchronization
